+++ small_text/integrations/pytorch/utils/data.py
@@ -51,15 +51,38 @@ def dataloader(data, batch_size, collate_fn, train=True):
                       sampler=sampler)
 
 
-# TODO: document that multi-label behaviour may be sub optimal
 def get_class_weights(y, num_classes, eps=1e-8):
+    """Computes class weights inversely proportional to their observed occurence. The most frequent
+    class receives a weight of `1.0` and the other classes receive a weight greater than that.
+
+    Parameters
+    ----------
+    y : np.ndarray[int] or csr_amtrix
+        A sequence of labels.
+    num_classes : int
+        The number of classes.
+    eps : float
+        A stability term which prevents division by zero.
+
+    Returns
+    -------
+    class_weights : torch.FloatTensor
+        A tensor of class weights.
+
+    Note
+    ----
+    This functionality is intended for single-label classification and is likely to be a
+    suboptimal solution for a multi-label scenario.
+    """
     label_counter = _get_class_histogram(y, num_classes, normalize=False)
-    pos_weight = torch.ones(num_classes, dtype=torch.float)
+    class_weights = torch.ones(num_classes, dtype=torch.float)
     num_samples = label_counter.sum()
     for c in range(num_classes):
-        pos_weight[c] = (num_samples - label_counter[c]) / (label_counter[c] + eps)
+        class_weights[c] = (num_samples - label_counter[c]) / (label_counter[c] + eps)
 
     if num_classes == 2:
-        pos_weight[pos_weight.argmin()] = 1.0
+        class_weights[class_weights.argmin()] = 1.0
+    else:
+        class_weights /= class_weights.min()
 
-    return pos_weight
+    return class_weights