+++ tests/integration/small_text/integrations/pytorch/classifiers/test_base.py
@@ -1,5 +1,10 @@
 import unittest
 
+import pytest
+import torch
+
+import numpy as np
+
 from small_text.integrations.pytorch.exceptions import PytorchNotFoundError
 
 try:
@@ -15,6 +20,11 @@ except PytorchNotFoundError:
 
 class PytorchClassifierImplementation(PytorchClassifier):
 
+    def __init__(self, num_classes, class_weight=None):
+        self.num_classes = num_classes
+        self.class_weight = class_weight
+        super().__init__()
+
     def _default_optimizer(self, base_lr):
         params = [param for param in self.model.parameters() if param.requires_grad]
         return params, AdamW(params, lr=base_lr, eps=1e-8)
@@ -29,6 +39,7 @@ class PytorchClassifierImplementation(PytorchClassifier):
         raise NotImplementedError()
 
 
+@pytest.mark.pytorch
 class PytorchClassifierTest(unittest.TestCase):
 
     def _get_dataset(self, num_samples=100, num_classes=4):
@@ -38,7 +49,7 @@ class PytorchClassifierTest(unittest.TestCase):
     def test_initialize_optimizer_and_scheduler_default(self):
         sub_train = random_text_classification_dataset(10)
 
-        classifier = PytorchClassifierImplementation()
+        classifier = PytorchClassifierImplementation(2)
         # initialize the model
         classifier.model = KimCNN(10, 60)
 
@@ -62,7 +73,7 @@ class PytorchClassifierTest(unittest.TestCase):
     def test_initialize_optimizer_and_scheduler_custom(self):
         sub_train = random_text_classification_dataset(10)
 
-        classifier = PytorchClassifierImplementation()
+        classifier = PytorchClassifierImplementation(2)
         # initialize the model
         classifier.model = KimCNN(10, 60)
 
@@ -84,3 +95,54 @@ class PytorchClassifierTest(unittest.TestCase):
 
         self.assertEqual(optimizer, optimizer_arg)
         self.assertEqual(scheduler, scheduler_arg)
+
+    def test_initialize_class_weights_binary_no_class_weights(self):
+        sub_train = random_text_classification_dataset(10)
+        classifier = PytorchClassifierImplementation(2)
+        class_weights = classifier.initialize_class_weights(sub_train)
+
+        self.assertIsNone(class_weights)
+
+    def test_initialize_class_weights_binary_balanced(self):
+        sub_train = random_text_classification_dataset(10)
+        sub_train.y = np.array([0] * 5 + [1] * 5)
+
+        classifier = PytorchClassifierImplementation(2, class_weight='balanced')
+        class_weights = classifier.initialize_class_weights(sub_train)
+
+        expected_weights = torch.ones(2, dtype=torch.float)
+
+        self.assertTrue(torch.equal(expected_weights.cpu(), class_weights.cpu()))
+
+    def test_initialize_class_weights_binary_imbalanced(self):
+        sub_train = random_text_classification_dataset(10)
+        sub_train.y = np.array([0] * 2 + [1] * 8)
+
+        classifier = PytorchClassifierImplementation(2, class_weight='balanced')
+        class_weights = classifier.initialize_class_weights(sub_train)
+
+        expected_weights = torch.FloatTensor([4.0, 1.0])
+
+        self.assertTrue(torch.equal(expected_weights.cpu(), class_weights.cpu()))
+
+    def test_initialize_class_weights_multi_class_balanced(self):
+        sub_train = random_text_classification_dataset(8)
+        sub_train.y = np.array([0] * 2 + [1] * 2 + [2] * 2 + [3] * 2)
+
+        classifier = PytorchClassifierImplementation(4, class_weight='balanced')
+        class_weights = classifier.initialize_class_weights(sub_train)
+
+        expected_weights = torch.ones(4, dtype=torch.float)
+
+        self.assertTrue(torch.equal(expected_weights.cpu(), class_weights.cpu()))
+
+    def test_initialize_class_weights_multi_class_imbalanced(self):
+        sub_train = random_text_classification_dataset(8)
+        sub_train.y = np.array([0] * 1 + [1] * 1 + [2] * 2 + [3] * 4)
+
+        classifier = PytorchClassifierImplementation(4, class_weight='balanced')
+        class_weights = classifier.initialize_class_weights(sub_train)
+
+        expected_weights = torch.FloatTensor([7.0, 7.0, 3.0, 1.0])
+
+        self.assertTrue(torch.equal(expected_weights.cpu(), class_weights.cpu()))